Generacion de documentacion para cada parte del codigo, y se agrego el diagrama de clases

+++ TrabajoFinal/personajes/Mago.java
@@ -3,17 +3,45 @@ package personajes;
 import armas.Arma;
 import java.util.Random;
 
+/**
+ * Clase concreta que representa a un personaje de tipo Mago.
+ * 
+ * El Mago posee una pasiva llamada "Poder arcano", que le permite tener mayor probabilidad de hacer un golpe critico
+ * 
+ * Esta clase hereda de {@link Personaje} y sobrescribe el método {@link #realizarAtaque(Personaje)}
+ * para implementar su pasiva
+ */
 public class Mago extends Personaje {
 
+    /**
+     * Constructor que inicializa al Mago con su nombre y arma.
+     *
+     * @param nombre Nombre del personaje.
+     * @param arma Arma que utilizará el Mago.
+     */
     public Mago(String nombre, Arma arma) {
         super(nombre, arma);
     }
 
+    /**
+     * Implementación de la pasiva del Mago.
+     * Imprime un mensaje indicando que se activa "Poder arcano".
+     */
     @Override
     public void aplicarPasiva() {
         System.out.println(nombre + " (Mago) activa su pasiva: Poder arcano.");
     }
 
+    /**
+     * Método sobrescrito para realizar el ataque del Mago.
+     * 
+     * Se disminuye el numero necesario para hacer un critico de 12 a 11 para que este tenga mayor probabilidad de
+     * Hacer que su ataque sea critico
+     * 
+     * Se imprime en consola el resultado de la tirada y el tipo de ataque realizado.
+     *
+     * @param oponente El personaje que recibe el ataque.
+     */
     @Override
     protected void realizarAtaque(Personaje oponente) {
         Random random = new Random();
@@ -32,7 +60,7 @@ public class Mago extends Personaje {
             } else {
                 System.out.println(nombre + " ataca con " + arma.getNombre());
             }
-            oponente.recibirDaño((int) daño);
+            oponente.recibirDaño(daño);
         }
     }
 }
